#104 Отдельный компонент для таблицы. Обобщение логики.
