+++ frontend/src/app/profile/components/profile-projects/profile-projects-chart/profile-projects-chart.component.scss
@@ -1,3 +1,4 @@
 :host {
   display: block;
+  margin-bottom: 20px;
 }
