+++ frontend/src/app/profile/components/profile-projects/profile-projects-table/profile-projects-table.component.html
@@ -1,0 +1,3 @@
+<p>
+  profile-projects-table works!
+</p>
