+++ frontend/src/app/profile/components/profile-projects/profile-projects.component.html
@@ -1,1 +1,3 @@
-<app-profile-projects-chart [projects]="projects"></app-profile-projects-chart>
+<app-profile-projects-chart [projects]="projects" [projectsMaxPeriod]="projectsMaxPeriod"></app-profile-projects-chart>
+
+<app-profile-projects-table [projects]="projects" [projectsMaxPeriod]="projectsMaxPeriod"></app-profile-projects-table>
