+++ frontend/src/app/profile/profile.module.ts
@@ -5,6 +5,7 @@ import { ProfileFormComponent } from './components/profile-form/profile-form.com
 import { ProfileHistoryComponent } from './components/profile-history/profile-history.component';
 import { ProfilePageComponent } from './components/profile-page.component';
 import { ProfileProjectsChartComponent } from './components/profile-projects/profile-projects-chart/profile-projects-chart.component';
+import { ProfileProjectsTableComponent } from './components/profile-projects/profile-projects-table/profile-projects-table.component';
 import { ProfileProjectsComponent } from './components/profile-projects/profile-projects.component';
 import { ProfileSubscriptionsComponent } from './components/profile-subscriptions/profile-subscriptions.component';
 
@@ -14,7 +15,8 @@ const components = [
   ProfileSubscriptionsComponent,
   ProfileFormComponent,
   ProfileProjectsComponent,
-  ProfileProjectsChartComponent
+  ProfileProjectsChartComponent,
+  ProfileProjectsTableComponent
 ];
 
 @NgModule({
